% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{.check_file_exists}
\alias{.check_file_exists}
\title{Verify a file path is non-NULL and exists}
\usage{
.check_file_exists(path)
}
\description{
Verify a file path is non-NULL and exists
}
\keyword{internal}
