% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-s3.R
\name{new_diagnostic_blocks}
\alias{new_diagnostic_blocks}
\title{Diagnostic-blocks data frame}
\usage{
new_diagnostic_blocks(x = data.frame(), component_cols = "")
}
\arguments{
\item{x}{A data frame.}

\item{component_cols}{Character; comma-joined list of detected component
columns (Morphologie / Immunphaenotypisierung / Zytogenetik /
Molekulargenetik). Carried as an attribute.}
}
\description{
S3 wrapper for the output of \code{\link[=zhn_prepare_diagnostic_blocks]{zhn_prepare_diagnostic_blocks()}}.
}
\seealso{
Other classes:
\code{\link[=new_cohort_df]{new_cohort_df()}},
\code{\link[=new_therapy_blocks]{new_therapy_blocks()}}
}
\concept{classes}
