% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-s3.R
\name{new_therapy_blocks}
\alias{new_therapy_blocks}
\title{Therapy-blocks data frame}
\usage{
new_therapy_blocks(x = data.frame(), patient_cols_used = "")
}
\arguments{
\item{x}{A data frame.}

\item{patient_cols_used}{Character; pipe-joined list of raw patient
columns coalesced into \code{patient} (carried as an attribute).}
}
\description{
S3 wrapper for the output of \code{\link[=zhn_prepare_therapy_blocks]{zhn_prepare_therapy_blocks()}}.
}
\seealso{
Other classes:
\code{\link[=new_cohort_df]{new_cohort_df()}},
\code{\link[=new_diagnostic_blocks]{new_diagnostic_blocks()}}
}
\concept{classes}
