% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{zhn_plot_cases_by_year}
\alias{zhn_plot_cases_by_year}
\title{Figure: cases per treatment year (vertical bar)}
\usage{
zhn_plot_cases_by_year(
  data,
  year_col = "behandlungsjahr",
  transparent = FALSE,
  xlab = "Jahr",
  ylab = "Anzahl"
)
}
\arguments{
\item{data}{A data frame with a treatment-year column.}

\item{year_col}{Name of the year column.}

\item{transparent}{Passed to \code{\link[=zhn_theme]{zhn_theme()}}.}

\item{xlab, ylab}{Axis labels.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Figure: cases per treatment year (vertical bar)
}
\seealso{
Other figures:
\code{\link{zhn_pal}},
\code{\link[=zhn_plot_diagnoses]{zhn_plot_diagnoses()}},
\code{\link[=zhn_plot_km]{zhn_plot_km()}},
\code{\link[=zhn_register_fonts]{zhn_register_fonts()}},
\code{\link[=zhn_save_plot]{zhn_save_plot()}},
\code{\link[=zhn_theme]{zhn_theme()}}
}
\concept{figures}
